moved stuff around, main file runs
