+++ node/go.mod
@@ -2,8 +2,10 @@ module node
 
 go 1.17
 
+require github.com/urfave/cli/v2 v2.3.0
+
 require (
-	github.com/golang/protobuf v1.5.2
-	google.golang.org/grpc v1.36.0
-	google.golang.org/protobuf v1.25.0
+	github.com/cpuguy83/go-md2man/v2 v2.0.0-20190314233015-f79a8a8ca69d // indirect
+	github.com/russross/blackfriday/v2 v2.0.1 // indirect
+	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
 )
